data_ingestion_to_vectordb/requirements.txt
opensearch-py==2.2.0
transformers==4.32.1
farm-haystack[aws]
farm-haystack[preprocessing]
farm-haystack[opensearch]
farm-haystack[inference]